requirements.txt
apiview==1.3.25
asgiref==3.7.2
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
colorama==0.4.6
coverage==7.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-import-export==3.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
py==1.11.0
PyJWT==2.8.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
